moved sha256 to utils
