+++ packages/js-sdk/src/sandbox/signature.ts
@@ -1,3 +1,5 @@
+import { sha256 } from '../utils'
+
 /**
  * Get the URL signature for the specified path, operation and user.
  *
@@ -16,24 +18,6 @@ interface SignatureOpts {
   user: string
   expirationInSeconds?: number
   envdAccessToken?: string
-}
-
-// Cross-platform crypto implementation
-async function sha256(data: string): Promise<string> {
-  // Use Node.js crypto if WebCrypto is not available
-  if (typeof crypto === 'undefined') {
-    // eslint-disable-next-line @typescript-eslint/no-var-requires
-    const { createHash } = require('node:crypto')
-    const hash = createHash('sha256').update(data, 'utf8').digest()
-    return hash.toString('base64')
-  }
-
-  // Use WebCrypto API if available
-  const encoder = new TextEncoder()
-  const dataBuffer = encoder.encode(data)
-  const hashBuffer = await crypto.subtle.digest('SHA-256', dataBuffer)
-  const hashArray = new Uint8Array(hashBuffer)
-  return btoa(String.fromCharCode(...hashArray))
 }
 
 export async function getSignature({
