+++ packages/js-sdk/src/utils.ts
@@ -1,0 +1,16 @@
+export async function sha256(data: string): Promise<string> {
+  // Use WebCrypto API if available
+  if (typeof crypto !== 'undefined') {
+    const encoder = new TextEncoder()
+    const dataBuffer = encoder.encode(data)
+    const hashBuffer = await crypto.subtle.digest('SHA-256', dataBuffer)
+    const hashArray = new Uint8Array(hashBuffer)
+    return btoa(String.fromCharCode(...hashArray))
+  }
+
+  // Use Node.js crypto if WebCrypto is not available
+  // eslint-disable-next-line @typescript-eslint/no-var-requires
+  const { createHash } = require('node:crypto')
+  const hash = createHash('sha256').update(data, 'utf8').digest()
+  return hash.toString('base64')
+}
